fix: rebase to master, adapting the vector store evaluation tests
